pyproject.toml
[project]
name = "support-agent"
version = "0.1.0"
description = "Whatsapp based support agent for customer query realted to products and details."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.1",
    "langchain-chroma>=1.1.0",
    "langchain-groq>=0.2.0",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.7",
    "langgraph-bigtool>=0.0.3",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langwatch-scenario>=0.7.15",
    "pandas>=3.0.0",
    "pip>=26.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
